Bugfixes, better names for language files, additions for phrase manager etc.

+++ viscacha/classes/function.global.php
@@ -145,6 +145,9 @@ function extract_dir($source, $realpath = true) {
 */
 function rmdirr($dirname) {
 	global $filesystem;
+	if (!file_exists($dirname)) {
+		return false;
+	}
 	if (is_file($dirname)) {
 		return $filesystem->unlink($dirname);
 	}
